int module/ -dryRun, -iter flag, before/after logging, added + minor changes

+++ src/go/intelligence/pipeline/delete_old_silents.go
@@ -21,7 +21,9 @@ func DeleteSilence(data <-chan models.AmJSON) <-chan models.AmJSON {
 	finalData := make(chan models.AmJSON)
 	go func() {
 		defer close(finalData)
-		deleteSilenceHelper()
+		if *utils.DryRun == false {
+			deleteSilenceHelper()
+		}
 		for each := range data {
 			finalData <- each
 		}
